--- src/components/AddTransaction.jsx
@@ -1,55 +1,0 @@
-import React, { useContext, useState } from 'react';
-import { GlobalContext } from '../context/GlobalState';
-import { generateID } from '../helpers';
-
-const AddTransaction = () => {
-    const {addTransaction} = useContext(GlobalContext);
-    const [text, setText] = useState('');
-    const [amount, setAmount] = useState(0);
-
-    const onSubmit = e => {
-        e.preventDefault();
-
-        const newTransaction = {
-            id: generateID(),
-            text,
-            amount: +amount
-        }
-
-        addTransaction(newTransaction);
-        setText('');
-        setAmount(0);
-    }
-
-    return (
-        <>
-            <h3>Add new transaction</h3>
-            <form onSubmit={onSubmit}>
-                <div className='form-control'>
-                    <label htmlFor='text'>Text</label>
-                    <input
-                        type='text'
-                        value={text}
-                        onChange={(e) => setText(e.target.value)}
-                        placeholder='Enter text...'
-                    />
-                </div>
-                <div className='form-control'>
-                    <label htmlFor='amount'>
-                        Amount <br />
-                        (negative - expense, positive - income)
-                    </label>
-                    <input
-                        type='number'
-                        value={amount}
-                        onChange={(e) => setAmount(e.target.value)}
-                        placeholder='Enter amount...'
-                    />
-                </div>
-                <button className='btn' disabled={!text || !amount}>Add transaction</button>
-            </form>
-        </>
-    );
-};
-
-export default AddTransaction;
