+++ src/components/expencesTracker/ChurchTax.jsx
@@ -1,18 +1,18 @@
 import React, { useContext } from "react";
-import { GlobalContext } from "../context/GlobalState";
+import { GlobalContext } from "../../context/GlobalState";
 
-const AnnualAmount = () => {
+const ChurchTax = () => {
   const { transactions } = useContext(GlobalContext);
   const amounts = transactions.map((transaction) => transaction.amount);
   const total = amounts.reduce((acc, item) => (acc += item), 0);
-  const annualAmount = (total * 12).toFixed(2);
+  const churchTax = ((total * 10) / 100).toFixed(2);
 
   return (
     <div className="count-inner-block">
-      <h4>Annual Amount</h4>
-      <h1>{annualAmount}</h1>
+      <h4>Church Tax - 10%</h4>
+      <h1>{churchTax}</h1>
     </div>
   );
 };
 
-export default AnnualAmount;
+export default ChurchTax;
