+++ src/components/expencesTracker/Transaction.jsx
@@ -1,0 +1,25 @@
+import React, { useContext } from "react";
+import { GlobalContext } from "../../context/GlobalState";
+
+const Transaction = ({ transaction }) => {
+  const { deleteTransaction } = useContext(GlobalContext);
+
+  const sign = transaction.amount < 0 ? "-" : "+";
+
+  return (
+    <li className={sign === "+" ? "plus" : "minus"}>
+      {transaction.text}{" "}
+      <span>
+        {sign}${Math.abs(transaction.amount)}
+      </span>
+      <button
+        className="delete-btn"
+        onClick={() => deleteTransaction(transaction.id)}
+      >
+        x
+      </button>
+    </li>
+  );
+};
+
+export default Transaction;
